Move MaterialApp home to separate widget
